Add initial support for configurable codecs
This commit adds the initial framework for allowing the configuration of
output codecs. There are three codec choices available:

* FLAC: (default) lossless, compression level 0-8
* OGG/Opus: lossy, bitrate 6-510kbps, Android 10+ only
* M4A/AAC: lossy, HE-AAC bitrate 24-32kbps, AAC-LC bitrate 32-128kbps

While all three codecs are implemented, RecorderThread is currently
hardcoded to use the default. The next step is to save and load the
codec configuration from SharedPreferences and the finally add the user
interface.

Issue: #21
Signed-off-by: Andrew Gunnerson <chillermillerlong@hotmail.com>