Rework buffering mechanism for recording
Previously, for passthrough recording, AudioRecord would record into a
100ms sized buffer (at the specified sample rate). For encoded
recording, it would record directly into MediaCodec's buffers, which may
or may not be sized appropriately.

This commit unifies the buffering scheme for both types of recording to
use a buffer twice the size of `AudioRecord.getMinBufferSize`. It also
introduces the `Encoder` abstraction to allow dealing with passthrough
and MediaCodec the same way.

Signed-off-by: Andrew Gunnerson <chillermillerlong@hotmail.com>