RecorderThread: Make AudioRecord internal buffer less excessively large
When the call ends, everything still remaining in AudioRecord's internal
buffer is lost because there's no way to read the remaining bytes after
recording is stopped. With 20x the buffer size on eg. a device with a
1920 byte minimum buffer size and a sample rate of 8000 Hz, this may
cause 2.4 seconds of audio at the end of the call to be lost.

This commit should not have a negative effect since the large buffer
size was meant to address buffer overruns and the root cause of that
problem was fixed in #72.

Signed-off-by: Andrew Gunnerson <chillermillerlong@hotmail.com>