Fix broken share action due to PendingIntent reuse
The intent for the "share" action is not distinguishable across
notifications when compared with Intent.filterEquals() because they
differ only in intent extras. This would cause the same PendingIntent to
be reused for every share action, which would reference an old URI.

This commit fixes the issue by setting the ACTION_SEND data to the same
value as the EXTRA_STREAM. The intent data is not used for this action,
but makes the intent sufficiently unique that a PendingIntent won't be
reused.

Signed-off-by: Andrew Gunnerson <chillermillerlong@hotmail.com>