Add support for pausing/resuming the recording
This commit implements basic pause/resume functionality via a
notification action in BCR's persistent notification. Pausing is
implemented by keeping the AudioRecord active, but discarding the data
instead of passing it to the encoder. Pausing/resuming does not happen
immediately, but after the processing of one buffer (which should be
under 50ms on most devices). It's fast enough to be imperceptible.

Issue: #198

Signed-off-by: Andrew Gunnerson <chillermillerlong@hotmail.com>