Switch to signing via SSH key
After this commit, all further commits, tags, and release artifacts will
be signed by a new SSH key:

    ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIDOe6/tBnO7xZhAWXRj3ApUYgn+XZ0wnQiXM8B7tPgv4

To authenticate this change, this commit is signed with the existing GPG
key (master key: 2233C479609BDCEC43BE9232F6A3B19090EFF32C, subkey:
19C3E6E8FB86AD288B38AF2714200B173D8A3BF5).

The reason for this change is to switch to stronger cryptography and
also to rely on simpler tools that are readily available. SSH is already
preinstalled on every desktop OS, including Windows.

Signed-off-by: Andrew Gunnerson <chillermillerlong@hotmail.com>