Add support for looking up the cached name in the call log
This commit adds a new `{call_log_name}` variable that can be used in
the filename template for including the call log cached name. The cache
name may be updated by the system dialer to include the name of the
business assocciated with the phone number.

This new variable is added to the default filename template, but only as
a last resort because it relies on polling for up to 2 seconds after the
call completes.

Fixes: #291

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>