Replace file retention slider with custom values only
This commit replaces the file retention slider with a simple dialog for
entering a custom value. Unlike the output format parameters, this
setting is purely a personal preference and there's no reason to suggest
hardcoded durations.

This was the last component that still used the Material 3 slider. No
more janky animations and abusing floating point values.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>