Add support for AMR-WB and AMR-NB
The patents have now expired everywhere in North America and Europe.

Note that this is not dumping the raw AMR call audio stream, which is
impossible on Android. The audio goes from AMR -> PCM -> AMR so the
output is a lossy reencoding. Users who want lossless recordings should
stick with FLAC or WAV.

Closes: #264

[1] https://voiceage.com/Patent-Portfolio.html
[2] https://voiceage.com/Patent-Portfolio-Essential.html
[3] https://patents.google.com/patent/EP1354315B1/es
[4] https://brevets-patents.ic.gc.ca/opic-cipo/cpd/eng/patent/2423651/summary.html

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>