Add support for direct boot
This allows BCR to record calls prior to the device being initially
unlocked after a reboot. In the BFU (before unlock state), recordings
are temporarily stored in an internal device-protected storage
directory. If the call completes before the initial unlock, then a
migration service that automatically runs after unlock will move the
files to the output directory. If the device is unlocked while the call
is still ongoing, then the recording will be moved to the output
directory at the end of the call.

There are some limitations, like not being able to look up contacts or
the call log, but most of BCR's will basically work as expected.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>