Add support for using contact groups in record rules
This is unfortunately a relatively big change because unlike single
contacts, Android does not provide a standard intent for apps to select
a contact group. We have to implement our own contact group picker UI.

Following our current convention, the most specific rule wins as a side
effect of how rules are sorted. If a call matches both a contact and a
contact group, the contact rule has higher precedence.

Fixes: #536

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>