Implement new record rules system
The new record rule system now supports matching the following
conditions:

* Call number (contact, contact group, unknown numbers)
* Call type (incoming, outgoing, conference)
* SIM slot

In addition, there is new action available that completely ignores a
call instead of just deferring deletion of the recording until the end
of the call.

Due to the increased flexibility, the record rules list UI no longer
sorts the rules, nor does it have a hardcoded unknown calls rule. The
only mandatory rule is the default rule. The rules can be freely
reordered by dragging them or removed by swiping them. Because these
gestures can't be performed via Talkback and other accessibility tools,
there are equivalent accessibility actions for moving a rule up in the
list, moving the rule down in the list, and removing a rule.

The new rules system is a superset of the old rules system, so all old
rules are automatically migrated.

Closes: #406
Closes: #509
Closes: #607

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>